added user auth to cleaning routes

+++ backend/app/db/repositories/cleanings.py
@@ -1,52 +1,45 @@
 from typing import List
 from uuid import UUID
 
-from fastapi import HTTPException, status
-
+from app.core.error import APIException
+from app.core.error_code import ErrorCode
+from app.models import User
 from app.models.cleaning import Cleaning
 from app.models.schemas.cleaning import CleaningBase, CleaningOut, CleaningUpdateIn
 
 
 class CleaningRepository:
-    async def get_all_cleanings(self) -> List[CleaningOut]:
-        return [CleaningOut.model_validate(cleaning) for cleaning in await Cleaning.all()]
+    async def get_all_cleanings(self, user: User) -> List[CleaningOut]:
+        return [CleaningOut.model_validate(cleaning) for cleaning in await Cleaning.filter(user_id=user.id)]
 
-    async def get_cleaning_by_id(self, cleaning_id: UUID) -> CleaningOut | None:
-        cleaning = await Cleaning.get_or_none(id=cleaning_id)
+    async def get_cleaning_by_id(self, cleaning_id: UUID, user: User) -> Cleaning:
+        cleaning = await Cleaning.get_or_none(id=cleaning_id, user_id=user.id)
         if cleaning is None:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail=f"Cleaning object not found",
-            )
+            raise APIException(ErrorCode.cleaning_not_found)
 
         return cleaning
 
-    async def create_cleaning(self, new_cleaning: CleaningBase) -> CleaningOut:
-        cleaning = await Cleaning.create(**new_cleaning.model_dump())
-        return CleaningOut.model_validate(cleaning)
+    async def create_cleaning(self, new_cleaning: CleaningBase, user: User) -> Cleaning:
+        return await Cleaning.create(**new_cleaning.model_dump(), user_id=user.id)
 
-    async def update_cleaning(self, cleaning_id: UUID, cleaning_in: CleaningUpdateIn) -> CleaningOut:
+    async def update_cleaning(self, cleaning_id: UUID, cleaning_in: CleaningUpdateIn, user: User) -> Cleaning:
         cleaning = await Cleaning.get_or_none(id=cleaning_id)
         if cleaning is None:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail=f"Cleaning object not found",
-            )
+            raise APIException(ErrorCode.cleaning_not_found)
+        if cleaning.user_id != user.id:
+            raise APIException(ErrorCode.cleaning_unauthorized_access)
 
-        cleaning_object = CleaningOut.model_validate(cleaning).model_dump()
-        for key, value in cleaning_in.model_dump().items():
-            if value is not None:
-                cleaning_object[key] = value
+        cleaning_dict = CleaningOut.model_validate(cleaning).model_dump()
+        cleaning_dict |= {key: value for key, value in cleaning_in.model_dump().items() if value is not None}
 
-        cleaning.update_from_dict(cleaning_object)
-        return CleaningOut.model_validate(cleaning)
+        await cleaning.update_from_dict(cleaning_dict)
+        return cleaning
 
-    async def delete_cleaning(self, cleaning_id: UUID):
+    async def delete_cleaning(self, cleaning_id: UUID, user: User):
         cleaning = await Cleaning.get_or_none(id=cleaning_id)
         if cleaning is None:
-            raise HTTPException(
-                status_code=status.HTTP_404_NOT_FOUND,
-                detail=f"Cleaning object not found",
-            )
+            raise APIException(ErrorCode.cleaning_not_found)
+        if cleaning.user.id != user.id:
+            raise APIException(ErrorCode.cleaning_unauthorized_access)
 
         await cleaning.delete()
